frontend: Refactor Indices page
- Fix editing indices
- Use default values when editing an index
- Add format validation
- Clean styles a little
